Fix for Boost 1.54 support
Use boost::make_tuple instead of std::make_tuple to produce composite keys.
Tested with Boost 1.54.0

Note this is expected to keep working on newer versions: "Everywhere where
std::tuples are used in the library interface, boost::tuples can be resorted to
in their place. The converse, however, is not true."
http://www.boost.org/doc/libs/1_63_0/libs/multi_index/doc/compiler_specifics.html