Add progress bar to dependency install scripts
For Travis builds, while building ceres and opencv, there are two conflicting
considerations:
    - `make` prints a lot of output not desired in the travis log
    - silencing `make` gives no indication of progress, and travis will time
      out without output (necessitating travis_wait)

To address this, pipe the script output to an awk script which prints a simple
progress bar (a dot for every 10 lines of make output). This means instead of
just "..." the script will print lots of dots. The rate may need tuning.

Use -j$(nproc) for make command -  works on Ubuntu and travis.