Move rotation from kinematics to geometry. (#51)
This commit move the rotation class from wave_kinematics to
wave_geometry, as that is a more sensible location for the
rotation and transformation class.  Wave kinematics is better
suited for motion models and other vehicle kinematics code.

The rotation class source and test files are moved to
wave_geometry, CMakeLists are updated accordingly, and
the documentation structure is updated as well.