Add landmark measurement container
Continue #41

The LandmarkMeasurement is a Measurement with an additional landmark_id
property, and a value type of Eigen::Vector2d representing pixel coordinates.

The LandmarkMeasurementContainer stores these measurements and offers these new
operations (on top of the MeasurementContainer methods):

- getLandmarkIDs()
- getLandmarkIDsInWindow()
- getTrack()
- getTrackInWindow()

Interpolation is not supported for landmark measurements.

I realize a lot of code is duplicated from the MeasurementContainer class
template. I started on an implementation that would reuse code, but because the
underlying types are different, code reuse requires an even more template-heavy
approach. I guessed reviewers would not be fans. I suggest pushing this
duplicated code and refactoring later.