Add range operations to measurement containers (#80)
Continuing #41

Add a number of methods working with iterators to MeasurementContainer, to
complete the interface and help with next work (dealing with ranges of landmarks).

New methods:
* getTimeWindow(): get all measurements between two times
* getAllFromSensor(): get all measurements from given sensor id
* constructor taking iterators: make new container with contents of a range
* erase taking iterators: erase measurements in a range
* insert taking iterators: insert contents of a range

These can be combined to, for example, copy a range of measurements to another
MeasurementConatiner or to a std::vector.

Also:
* refactor get() interpolation implementation

Previously, the container was sorted an odd way (sensor id first, then time)
which made this method awkward.  Refactor to sort by time first, then sensor
id, and implement the "complex search" method described at
http://www.boost.org/doc/libs/1_63_0/libs/multi_index/doc/examples.html#example6
for get().