Use Eigen::Map instead of Ref to fix issue
The Ref was pointing at the wrong data. Mysterious, still investigating..
Anyway, map is more straightforward.

The ugliness of the ValueView derived classes is to be addressed later ("todo
add more syntactic sugar")

Also remove some redundant constuctors and add virtual destructor to be clear
that ValueView is a base class.

The ugliness of the ValueView