Remove unneeded typedef
The class definition is organized, top to bottom:

- typedefs
- methods
- data members

The first typedef is private, and is first because it is used for convenience of defining the other types.

Remove an unneeded typedef used only in the subsequent two lines.
Change "protected" traits typedef to private, and redefine it in derived classes to avoid possibly confusing "Base::traits".