Add PerfectPrior class
(Squashed commit)

- Add addPerfectPrior

- Simplify constructors

When we have to copy a value, accept a value.

The alternative is to have two overloads: (T&&) and (T const&) but this adds repeated code.

See https://stackoverflow.com/questions/7592630/is-pass-by-value-a-reasonable-default-in-c11

- Refactor perfect priors

- Setting variables constant is handled in graph_wrapper, not by factors
- Remove ZeroNoise type, instead, use "void" NoiseType
- evaluateRaw calls a different templated helper function if the Factor has zero noise

- Add PerfectPrior class

Remove the evaluateSpecialized helper, too complicated.

Have addPerfectPrior add a special PerfectPrior class. This class does nothing on evaluateRaw.

Set the variable value, as well setting it constant. This is the key!