Export with Wave:: namespace instead of wave_ prefix
This means users of the library will be able to link to components such as
Wave::utils instead of Wave::wave_utils or just wave_utils.

The benefit of using :: is CMake recognizes it as a namespace, and actually
checks that the name exists when calling target_link_libaries.

This change involved:
 * Add a WAVE_ADD_LIBRARY function, which is called instead of ADD_LIBRARY for
   each of our modules
   - It combines the typical ADD_LIBARY, INSTALL, and some other calls
 * For each library, set EXPORT_NAME property to its name stripped of wave_
   (for example, wave_utils is exported as utils, which becomes Wave::utils)
 * For each library, add Wave::utils alias, so internal links can use it too
 * When the whole package is exported, the "Wave::" namespace is specified

An alternative would be to remove the prefix entirely from each project. Then
the target would actually be called "utils", etc. Setting OUTPUT_NAME would
keep the prefix in the library file name (e.g. libwave_utils.so). I chose not
to go this route for now, as it might be more disruptive, though using regex to
remove the prefix seems less elegant.