Fix missing compile flag from VTK
VTK_REQUIRED_CXX_FLAGS is set to -Wno-deprecated (important on travis since
it has Werror). In the normal course of thing this seems to get added directly
to CMAKE_CXX_FLAGS, but I put FindPCL into a new scope specifically to avoid
things like that.

Explicitly add the flags to the target properties.