Add benchmarks for MeasurementContainer emplace() and get()
Add benchmarks which time emplace() and get() functions with a simple
measurement value type, as well as the equivalent functions for a bare-bones
boost multi_index_container.

These benchmarks show that emplace() is working fine, but get() is terribly
inefficient compared to the bare-bones find(), even when no interpolation is
done.