CMake: allow partial build and finding components (#231)
Support builds which do not include all modules.

- There is a BUILD_wave_ option for each module, which the user can set OFF to disable that module
- Each module checks its dependencies and disables its own build if any are missing
There is a new helper, WAVE_ADD_MODULE, which does almost everything
- The user can call find_package(wave COMPONENTS geometry ...)
- Each module's dependencies are checked again, in case a dependency present at build time is no longer found.

This dependency system is less sophisticated than those used by some large projects. For example, a module can't be disabled by default, the modules must be listed in top-level CMakeLists in a particular order with no circular dependencies allowed, and the messages could be prettier. However it is simple and enough for now.

Also:

* CMake: don't print message for each test target

It's not really useful, gets out of hand with many tests

* Refactor CMake: add WAVE_ADD_MODULE helper

* Allow find_package(wave COMPONENTS ...)

* Update ReadMe

* Use Release build type by default

I accidentally removed setting CMAKE_BUILD_TYPE from wave_matching and now it's
really slow. But the build type should probably be the same across different modules, and
should be overridable.