Disable strict-overflow warning
This warning warns that the compiler assumed (N + a) > N is true by assuming N
+ a does not overflow. It is an optimization-level-dependent warning.

Since signed overflow is undefined behaviour anyway, I don't see this as very
useful.

There's also not much we can do in this case since N+a is added outside of a
function, and the comparison is inside.

See https://stackoverflow.com/q/12984861/

Also use the cmake3 command add_compile_options()