Replace wave_geometry with submodule (#291)
* Add wave_geometry as a submodule
The library at https://github.com/wavelab/wave_geometry (which can also be installed separately) will be built as a component (aka "module") of libwave

* Remove the old module (also called "wave_geometry")

* Update dependent code in wave_benchmark
Just quick changes to keep the same behaviour for now. It remains to integrate new geometry library more fully.

Make some changes to CMake:

* Change how gtest is found by adding AddGtest.cmake script
* Add GTest::Main target to match wave_geometry
  (Pave the way to remove bundled gtest in future, if desired)

* Install Find*.cmake modules
Removes inconsistency in which Eigen3 is found during build of libwave and
during later use of libwave via find_package.

Note most of the CMake that allows wave_geometry to be built either standalone or as a component of libwave is in wave_geometry itself.